selection cuts applied in different script
