+++ devel/external/AEGIS/get_in_galsim.py
@@ -55,45 +55,6 @@ from astropy.io import fits
 from astropy.table import Table,Column, vstack, hstack, join
 import os,glob
 
-def apply_selection(args, all_seg_ids):
-    """Applies selection cuts to galaxies in each segemnt.
-    Selection criteria are: 
-    1) Closest masked pixel is more than 11 pixels away
-                  OR 
-    2) Average flux in 9*9 pixels centered on the closest replaced pixel (before masking) is less than 
-       half of the peak flux of central galaxy
-                   AND  
-    3) SNR in elliptical gaussian filter is positve.
-    
-    Galaxies statifying these criteria in all filters are selected for the final catalog.
-    """
-    for seg in all_seg_ids:
-        cat={}
-        temp=[]
-        for f, filt in enumerate(args.filter_names):
-            cat_name = args.main_path + seg + '/' + filt + '_with_pstamp.fits'
-            cat[f] = Table.read(cat_name, format= 'fits')
-            #Select only good postage stamps
-            cond1 = cat[f]['min_mask_dist_pixels'] > 11
-            cond2 = cat[f]['average_mask_adjacent_pixel_count']/ cat[f]['peak_image_pixel_count'] < 0.5
-            cond3 = cat[f]['sn_ellip_gauss'] > 0
-            good, = np.where((cond1 | cond2) & cond3)
-            temp.append(good)
-        if args.apply_cuts is True:
-            select_gal = reduce(np.intersect1d,temp)
-        else:
-            # If no selection cut has to be applied
-            select_gal = range(len(cat[f]))
-        for f, filt in enumerate(args.filter_names):
-            new_cat_name = args.main_path + seg + '/' + filt + '_selected.fits'
-            new_cat = cat[f][select_gal]
-            if os.path.isfile(new_cat_name) is True:
-                subprocess.call(["rm", new_cat_name]) 
-            print "Catalog with pstamps saved at ", new_cat_name
-            new_cat.write(new_cat_name, format='fits') 
-        select_gal_list= args.main_path + seg + '/gal_in_cat.txt' 
-        np.savetxt(select_gal_list, select_gal)
-
 def assign_num(args, all_seg_ids):
     """Assigns individual identification number to each object"""
     seed =122
@@ -292,7 +253,7 @@ def get_fits_catalog(args, index_table, all_seg_ids):
         print "Savings fits file at ", path + file_name
         final_table[ord_indx].write(path + file_name, format='fits',
                                                 overwrite=True)
-def get_in_galsim(args):
+def make_all(args, input_cat):
     """Saves Postage stamps and final catalogs in a format that can be read by
      galsim modules"""
     if os.path.isdir(args.main_path + args.out_dir) is False:
@@ -313,37 +274,4 @@ def get_in_galsim(args):
     get_fits_catalog(args, idx,
                      all_seg_ids)
 
-if __name__ == '__main__':
-    import subprocess
-    import numpy as np
-    from argparse import ArgumentParser
-    parser = ArgumentParser()
-    parser.add_argument('--n_filters', type=int, default=2,
-                        help="number of image filters [Default: 2]")
-    parser.add_argument('--filter_names', default= ['f606w','f814w'],
-                        help="names of filters [Default: ['f606w','f814w']]")
-    parser.add_argument('--main_path',
-                        default = '/nfs/slac/g/ki/ki19/deuce/AEGIS/AEGIS_catalog_full/')
-    parser.add_argument('--out_dir', default = "AEGIS_training_sample/",
-                        help="directory containing the final catalog")
-    parser.add_argument('--cat_name', default = "AEGIS_galaxy_catalog_filter_25.2.fits",
-                        help="Final catalog name")
-    parser.add_argument('--gal_im_name', default = "AEGIS_galaxy_images_filter_25.2_number.fits",
-                        help="Final name of galaxy images")
-    parser.add_argument('--psf_im_name', default = "AEGIS_galaxy_PSF_images_filter_25.2_number.fits",
-                        help="Final name of PSF images")
-    parser.add_argument('--selec_file_name', default = "AEGIS_galaxy_catalog_filter_25.2_selection.fits",
-                        help="Catalog with selection information")
-    parser.add_argument('--noise_file_name', default = "acs_filter_unrot_sci_cf.fits",
-                        help="File with correlation function of noise")
-    parser.add_argument('--file_filter_name', default =['V', 'I'] ,
-                        help="Name of filter to use ")
-    parser.add_argument('--fits_file_name', default = "AEGIS_galaxy_catalog_filter_25.2_fits.fits",
-                        help="Name of Catalog with fit information")
-    parser.add_argument('--seg_list_file', default ='/nfs/slac/g/ki/ki19/deuce/AEGIS/unzip/seg_ids.txt',
-                        help="file with all seg id names" )
-    parser.add_argument('--apply_cuts', default = True,
-                        help="Remove galaxies with imperfect masking during pstamp cleaning step.[Default=True]")
-    args = parser.parse_args()
-    get_in_galsim(args)
