+++ devel/external/AEGIS/run_batch_third.py
@@ -8,7 +8,7 @@ def run_clean_pstamps():
     for fl in glob.glob('outfile/out_3_seg_*'):
         os.remove(fl)
     file_name ='/nfs/slac/g/ki/ki19/deuce/AEGIS/unzip/seg_ids.txt'
-    all_seg_ids = np.loadtxt(file_name, delimiter=" ",dtype='S2')[48:63]
+    all_seg_ids = np.loadtxt(file_name, delimiter=" ",dtype='S2')
     for seg_id in all_seg_ids:
         print 'SEG ID ', seg_id
         outfile = 'outfile/out_3_seg_{0}.txt'.format(seg_id)
